+++ src/Microsoft.DotNet.Interactive.Formatting/TypeFormatterSourceAttribute.cs
@@ -8,11 +8,11 @@ namespace Microsoft.DotNet.Interactive.Formatting
     [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, AllowMultiple = true, Inherited = true)]
     public class TypeFormatterSourceAttribute : Attribute
     {
-        public Type FormatterSourceType { get; }
-
         public TypeFormatterSourceAttribute(Type formatterSourceType)
         {
             FormatterSourceType = formatterSourceType;
         }
+
+        public Type FormatterSourceType { get; }
     }
 }
